HTML/CSS has a basic game board (#board), that will house game tiles. Each tile will containing either nothing, an 'X', or an 'O'. Upon initializing, the board and its tiles will all be blank.
JavaScript module patterns created - gamePlay: will house methods that will control the flow of the game ; gameBoard: will represent the virtual display of the board, with a method that successfully renders the current mock-state of the board (held in the currentBoard property) in the user's browser. The currentBoard property holds the current pieces and state of the board, this will be used for reference to change tiles from blanks to 'X's or 'O's and will be used for game logic.