gamePlay module object contains a property, roundCount, that will store the current round of the game. This will be used in the gameLogic module object to determine which players turn it is and what marker to assign/print to the boardTiles that a player clicks on.
gameLogic module object contains a function, whosTurnIsIt, that will evaluate which player's turn it is based on the gamePlay.roundCount number, and then return the appropriate player index who's turn it is from the gamePlay.players array and then increment the gamePlay.roundCount by 1. This function will  be used as an event handler for the boardTiles on the game board that players 'click' on, adding the player's marker of which player turn it is.